🚚(frontend) move converse.ts in new folder and plugins in specific files
In addition to these changes, i also removed from the dom the automatically
injected converse UI chat, and updated all filepaths to converse.ts
