+++ src/frontend/components/Chat/index.tsx
@@ -1,9 +1,9 @@
 import { Box, BoxExtendedProps } from 'grommet';
 import React, { useEffect } from 'react';
 
-import { useVideo } from '../../data/stores/useVideo';
-import { Video } from '../../types/tracks';
-import { converseMounter } from '../../utils/converse';
+import { useVideo } from 'data/stores/useVideo';
+import { Video } from 'types/tracks';
+import { converseMounter } from 'utils/conversejs/converse';
 
 interface ChatProps {
   video: Video;
